Move s3_to_dc_v2 to indexing_tools modue
