+++ setup.cfg
@@ -73,7 +73,7 @@ console_scripts =
     download-esa-worldcereal-cogs = deafrica.data.esa_worldcereal:download_esa_worldcereal_cogs
     create-esa-worldcereal-stac = deafrica.data.esa_worldcereal:create_esa_worldcereal_stac
     create-wapor-v3-stac = deafrica.data.wapor_v3:create_wapor_v3_stac
-    s3-to-dc-v2 = deafrica.data.s3_to_dc_v2:cli
+    s3-to-dc-v2 = deafrica.data.indexing_tools.s3_to_dc_v2:cli
 
 [options.packages.find]
 include =
